# Change 2-way diff logic to use a separate shared object.