This removes the need for two objects per DOM node (`attrs` and `newAttrs`) in addition to `attrsArr`.

There shouldn't be much of a performance difference either way since attributes should generally line up as before. There is a slight benefit to initial creation time due to the extra objects not being created.