@jridgewell 2016-09-30
LGTM. Has the been perf tested?