@sparhami 2016-09-30
On the list test, if you comment out the lining up flow for selection (and leaving it in for creation), I get:

| test | old diff | new diff |
|------|-----------|------------|
| selection | 0.5ms | 1.1ms|
| creation | 4.6ms | 4.2ms |

My thinking is that the attributes not lining up should not happen too often. The performance of the second part mattered more before we had the logic to do the whole diff if they line up (/ for newly created elements) in the array check. I think the performance benefit to creation, simplicity and reduced memory usage probably outweigh the extra diff time if you hit that flow.