@jridgewell 2016-10-04
I wonder if we could get the same creation speed boost by lazy creating the `diffAttrs` on each node. I think the selection slowdown is coming from iterating an object that's been de-opt'd (the `delete`).