@sparhami 2016-10-04
> I wonder if we could get the same creation speed boost by lazy creating the diffAttrs on each node. I think the selection slowdown is coming from iterating an object that's been de-opt'd (the delete).

I think deoptimizing the object here should be okay. They will be treated as generic maps, which is really what the are. This should allow V8 not to need to keep around a class object for each of the different types it encounters. It also should be a pretty rare flow. If you are transforming one subree into another, you'll probably also be creating DOM nodes which should dominate over the attributes diffing time.

The tests don't have too may different types, but in practice this could end up saving memory by not needing a bunch of hidden classes.

A few things to look into are:

1. What about just using Map when available and using an Object as a fallback?
2. What are the savings by forcing this into a generic map in terms of memory usage?
  - Probably need to create a test that creates a bunch of different object types.
3. Can we force it into a generic map up front (by doing a delete inside createMap), this should save V8 a bunch of time creating all the hidden classes, then killing them later. This is probably really hard to measure in practice.