@sparhami 2016-10-06
Looking into the memory savings, based on heap dumps of before/after for the list test, I see a reduction in size for all the `NodeData` from 77K bytes bytes to 52K bytes. The list test doesn't have too many attributes per DOM node and that is without the `newAttrs` objects ever being touched.

Looking into the performance, I tried:

```
  if (i < arguments.length || j < attrsArr.length) {
    const attrsStart = j;
    const used = {};
    const prevAttrsMap = {};

    for (; j < attrsArr.length; j += 2) {
      prevAttrsMap[attrsArr[j]] = attrsArr[j + 1];
    }

    for (j = attrsStart; i < arguments.length; i += 2, j += 2) {
      const name = arguments[i];
      const value = arguments[i + 1];

      if (prevAttrsMap[name] !== value) {
        updateAttribute(node, name, value);
      }

      attrsArr[j] = name;
      attrsArr[j + 1] = value;

      prevAttrsMap[name] = used;
    }

    if (j < attrsArr.length) {
      attrsArr.length = j;
    }

    /*
     * At this point, only have attributes that were present before, but have
     * been removed.
     */
    for (const name in prevAttrsMap) {
      if (prevAttrsMap[name] !== used) {
        updateAttribute(node, name, undefined);
      }
    }
  }
```

which has about the same performance as before. So what seems to be going on is that the code is really hot (due to collecting lots of samples in a tight loop) where there aren't many different object types. This does not really reflect real application behavior since you will have lots of different object shapes hitting the function in practice. To account for this, I added `elementOpen` calls to force a bunch of different object shapes to hit the diffing logic in each test loop:

```
...
elementClose('ul');

elementOpen('div');
  for (var i = 0; i < 2000; i += 1) {
    elementOpen('span', null, null,
        'data-one' + i, '',
        'data-two' + i, '',
        'data-three' + i, '');
    elementClose('span');
  }
elementClose('div');
```

and then testing master, this PR and the modification above. With the above code, the PR performed the best, followed by master and the modification above performing about the same. So for real world usage, this PR could improve the diffing speed for the two-way diff as well (although this is not a critical part of the overall performance). If we do add a test for the two-way diff portion of the code, we will want to make sure that we include some logic like the above to make sure we are accurately capturing what would happen in an actual application.