@jridgewell 2016-10-06
This seems good to me. The slowdown only happens in the slow path any way, but the memory savings are pretty damn good.